+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# @addfox/common
 
+## 0.1.1-beta.11
+
+### Patch Changes
+
+- chore: update manifest function
+
 ## 0.1.1-beta.10
 
 ### Patch Changes