+++ packages/plugins/rsbuild-plugin-extension-monitor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @addfox/rsbuild-plugin-extension-monitor
 
+## 0.1.1-beta.11
+
+### Patch Changes
+
+- chore: update manifest function
+- Updated dependencies
+  - @addfox/common@0.1.1-beta.11
+  - @addfox/core@0.1.1-beta.11
+
 ## 0.1.1-beta.10
 
 ### Patch Changes